feat: only admin or team manager (of this team) can have actions button
This commit adds the ability for admins or managers to edit and delete seasons in the team page. The columns array is updated to include the 'actions' column only for users with the role of admin or manager. This change improves the functionality and user experience of the team page.

Co-authored-by: NhiNguyen34  <yennhi.nguyenngoc03@gmail.com>