chore: add disabled state for team select in user update page
This commit adds a disabled state for the team select component in the user update page. The team select component will be disabled for users with a role other than 1 (admin). Additionally, a message is displayed for users with a role of 0 (non-admin) to contact the admin to change their team.

Co-authored-by: 21522173huy  <tulehuy.003@gmail.com@gmail.com>